Rename Button.vue to ButtonSubmit.vue
